Fixed the bug that the websocket will disconnect when user switches view in collaboration frontend
